requirements.txt
agate==1.7.1
aniso8601==9.0.1
attrs==23.1.0
Babel==2.13.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
db-dtypes==1.1.1
dbt-bigquery==1.7.2
dbt-core==1.7.2
dbt-extractor==0.5.0
dbt-semantic-interfaces==0.4.1
Flask==3.0.0
Flask-Cors==4.0.0
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-bigquery==3.13.0
google-cloud-core==2.3.3
google-cloud-dataproc==5.7.0
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
grpc-google-iam-v1==0.12.7
grpcio==1.59.3
grpcio-status==1.59.3
idna==3.4
importlib-metadata==6.8.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
leather==0.3.4
Logbook==1.5.3
MarkupSafe==2.1.3
mashumaro==3.10
minimal-snowplow-tracker==0.0.2
more-itertools==10.1.0
msgpack==1.0.7
networkx==3.2.1
numpy==1.26.2
packaging==23.2
pandas==2.1.3
parsedatetime==2.6
pathspec==0.11.2
proto-plus==1.22.3
protobuf==4.25.1
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.13
python-dateutil==2.8.2
python-slugify==8.0.1
pytimeparse==1.1.8
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
requests==2.31.0
rpds-py==0.13.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
Werkzeug==3.0.1
zipp==3.17.0
